refactor: centralize theme storage helpers
